+++ .dockerignore
@@ -1,0 +1,5 @@
+node_modules
+.git
+.gitignore
+*.md
+dist
